+++ src/Vector/vector_std_pack_unpack.ipp
@@ -33,7 +33,7 @@
 			Packer<size_t, Memory1>::pack(mem,obj.size(),sts);
 
 			//Call a packer in nested way
-			for (int i = 0; i < obj.size(); i++) {
+			for (size_t i = 0; i < obj.size(); i++) {
 				obj.get(i).template pack<prp...>(mem,sts);
 			}
 		}
@@ -77,7 +77,7 @@
 			//Call an unpacker in nested way
 
 			//std::cout<< demangle(typeid(obj.get(1)).name()) << std::endl;
-			for (int i = 0; i < obj.size(); i++) {
+			for (size_t i = 0; i < obj.size(); i++) {
 				obj.get(i).template unpack<prp...>(mem,ps);
 			}
 		}
@@ -117,7 +117,7 @@
 			v.push_back(sizeof(obj.size()));
 
 			//Call an packRequest in nested way
-			for (int i = 0; i < obj.size(); i++)
+			for (size_t i = 0; i < obj.size(); i++)
 			{
 				obj.get(i).template packRequest<prp...>(v);
 			}
@@ -150,7 +150,7 @@
 #endif
 			size_t res = 0;
 
-			for (int i = 0; i < n; i++) {
+			for (size_t i = 0; i < n; i++) {
 				res += obj.get(i).template packMem<prp...>(obj.get(i).size(),0);
 			}
